Fix dedicated server crash
