+++ fabric/src/main/java/com/lx862/jcm/mod/block/KCREnquiryMachineWall.java
@@ -4,7 +4,7 @@ import com.lx862.jcm.mod.block.base.WallAttachedBlock;
 import com.lx862.jcm.mod.block.behavior.EnquiryMachineBehavior;
 import com.lx862.jcm.mod.data.EnquiryLog;
 import com.lx862.jcm.mod.data.Entry;
-import com.lx862.jcm.mod.network.block.EnquiryUpdateGUIPacket;
+import com.lx862.jcm.mod.network.gui.EnquiryUpdateGUIPacket;
 import com.lx862.jcm.mod.registry.Networking;
 import com.lx862.jcm.mod.util.BlockUtil;
 import com.lx862.jcm.mod.util.VoxelUtil;
