+++ fabric/src/main/java/com/lx862/jcm/mod/network/gui/ClientHelper.java
@@ -1,10 +1,17 @@
 package com.lx862.jcm.mod.network.gui;
 
+import com.lx862.jcm.mod.data.Entry;
 import com.lx862.jcm.mod.render.gui.screen.*;
 import org.mtr.mapping.holder.BlockPos;
 import org.mtr.mapping.holder.MinecraftClient;
 import org.mtr.mapping.holder.Screen;
 
+import java.util.List;
+
+/**
+ * Because Network Packet class is shared between Client and Server, they must not contain any client code
+ * This serves as an abstraction to perform client-specific operation.
+ */
 public final class ClientHelper {
 
     public static void openButterflyLightScreen(BlockPos blockPos, int secondsToBlink) {
@@ -25,5 +32,13 @@ public final class ClientHelper {
 
 	public static void openSubsidyMachineGUIScreen(BlockPos blockPos, int pricePerUse,int cooldown) {
 		MinecraftClient.getInstance().openScreen(new Screen(new SubsidyMachineScreen(blockPos, pricePerUse, cooldown)));
+	}
+
+	public static void openEnquiryScreen(List<Entry> entries) {
+		MinecraftClient.getInstance().openScreen(new Screen(new EnquiryScreen(entries)));
+	}
+
+	public static void openRVEnquiryScreen(List<Entry> entries) {
+		MinecraftClient.getInstance().openScreen(new Screen(new RVEnquiryScreen(entries)));
 	}
 }
